+++ model/pdata/logs_alias.go
@@ -17,7 +17,6 @@ package pdata // import "go.opentelemetry.io/collector/model/pdata"
 // This file contains aliases for log data structures.
 
 import (
-	"go.opentelemetry.io/collector/model/internal"
 	"go.opentelemetry.io/collector/model/internal/pdata"
 )
 
@@ -35,11 +34,6 @@ type Logs = pdata.Logs
 
 // NewLogs is an alias for a function to create new Logs.
 var NewLogs = pdata.NewLogs
-
-// Deprecated: [v0.47.0] will be removed soon, only used internally.
-func LogsFromInternalRep(internal.LogsWrapper) Logs {
-	return Logs{}
-}
 
 // SeverityNumber is an alias for pdata.SeverityNumber type.
 type SeverityNumber = pdata.SeverityNumber