+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 🛑 Breaking changes 🛑
 
+- Remove pdata `InternalRep` deprecated funcs (#5018)
 - Remove service/defaultcomponents deprecated package (#5019)
 
 ### 🚩 Deprecations 🚩